+++ changelog.md
@@ -1,3 +1,15 @@
+# 3.13.2
+
+## Changes
+
+More info frames is no longer clamped to screen.
+
+## Bugfixes
+
+- *Selecting an item for "Award Later" during a session will now properly count the item as awarded.*
+- *Tier token items are now properly registered as such for usage in more info tooltips.*
+- *BoE's are recognized once again.*
+
 # 3.13.1
 
 ## Changes